+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/editTasks/editCdcSinkTask/EditCdcSinkTask.scss
@@ -7,20 +7,23 @@
         gap: sizes.$gutter-xxs;
         align-items: center;
 
+        --input-height: 30px;
+        --input-padding: 2px;
+
         .form-control {
-            min-height: 30px;
-            font-size: 14px;
+            min-height: var(--input-height);
+            font-size: sizes.$font-size-root;
         }
 
         .react-select-container {
-            font-size: 14px;
+            font-size: sizes.$font-size-root;
 
             .react-select__control {
-                min-height: 30px;
+                min-height: var(--input-height);
             }
 
             .react-select__value-container {
-                min-height: 28px;
+                min-height: calc(var(--input-height) - var(--input-padding));
                 padding-top: 0;
                 padding-bottom: 0;
             }
@@ -31,7 +34,7 @@
             }
 
             .react-select__indicators {
-                height: 28px;
+                height: calc(var(--input-height) - var(--input-padding));
             }
 
             .react-select__dropdown-indicator {
