+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/editTasks/editCdcSinkTask/sections/tables/tableEditor/EditCdcSinkTaskFieldMapping.tsx
@@ -1,0 +1,82 @@
+import { EmptySet } from "components/common/EmptySet";
+import { FormInput, FormSelect } from "components/common/Form";
+import { Icon } from "components/common/Icon";
+import { SelectOption } from "components/common/select/Select";
+import {
+    RootTablePath,
+    EmbeddedTablePath,
+} from "components/pages/database/tasks/ongoingTasks/editTasks/editCdcSinkTask/utils/editCdcSinkTaskFormPaths";
+import { EditCdcSinkTaskFormData } from "components/pages/database/tasks/ongoingTasks/editTasks/editCdcSinkTask/utils/editCdcSinkTaskValidation";
+import Button from "react-bootstrap/Button";
+import { useFieldArray, useFormContext } from "react-hook-form";
+
+type CdcColumnType = Raven.Client.Documents.Operations.CdcSink.CdcColumnType;
+type FieldMappingPath = RootTablePath | EmbeddedTablePath;
+
+export default function EditCdcSinkTaskFieldMapping({ path }: { path: FieldMappingPath }) {
+    const { control } = useFormContext<EditCdcSinkTaskFormData>();
+
+    const columnsFieldArray = useFieldArray({
+        control,
+        name: `${path}.columns`,
+    });
+
+    return (
+        <>
+            <div className="hstack justify-content-between mb-1">
+                <div>Field mapping</div>
+                <Button
+                    variant="link"
+                    size="sm"
+                    onClick={() => columnsFieldArray.append({ column: "", name: "", type: "Default" })}
+                >
+                    <Icon icon="plus" />
+                    Add field mapping
+                </Button>
+            </div>
+            {columnsFieldArray.fields.length === 0 ? (
+                <div className="panel-bg-2 p-1 rounded border border-secondary hstack justify-content-center">
+                    <EmptySet compact>No field mappings defined.</EmptySet>
+                </div>
+            ) : (
+                <div>
+                    <div className="field-mapping-row panel-bg-2 rounded-top p-1">
+                        <div>Source column</div>
+                        <div></div>
+                        <div>Target column</div>
+                        <div>Type</div>
+                        <div></div>
+                    </div>
+                    <div className="bg-body rounded-bottom p-1 vstack gap-1">
+                        {columnsFieldArray.fields.map((field, idx) => (
+                            <div key={field.id} className="field-mapping-row">
+                                <FormInput type="text" control={control} name={`${path}.columns.${idx}.column`} />
+                                <Icon icon="arrow-thin-right" margin="m-0" />
+                                <FormInput type="text" control={control} name={`${path}.columns.${idx}.name`} />
+                                <FormSelect
+                                    control={control}
+                                    name={`${path}.columns.${idx}.type`}
+                                    options={columnTypeOptions}
+                                />
+                                <Button
+                                    variant="link"
+                                    className="text-danger"
+                                    size="sm"
+                                    onClick={() => columnsFieldArray.remove(idx)}
+                                >
+                                    <Icon icon="trash" margin="m-0" />
+                                </Button>
+                            </div>
+                        ))}
+                    </div>
+                </div>
+            )}
+        </>
+    );
+}
+
+const columnTypeOptions: SelectOption<CdcColumnType>[] = [
+    { value: "Default", label: "Default" },
+    { value: "Json", label: "JSON" },
+    { value: "Attachment", label: "Attachment" },
+];
