+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/editTasks/editCdcSinkTask/sections/tables/tableEditor/EditCdcSinkTaskLinkedTableEditor.tsx
@@ -1,0 +1,37 @@
+import { FormGroup, FormInput, FormLabel } from "components/common/Form";
+import { LinkedTablePath } from "components/pages/database/tasks/ongoingTasks/editTasks/editCdcSinkTask/utils/editCdcSinkTaskFormPaths";
+import { EditCdcSinkTaskFormData } from "components/pages/database/tasks/ongoingTasks/editTasks/editCdcSinkTask/utils/editCdcSinkTaskValidation";
+import { useFormContext } from "react-hook-form";
+import EditCdcSinkTaskStringValueListEditor from "./EditCdcSinkTaskStringValueList";
+
+export default function EditCdcSinkTaskLinkedTableEditor({ path }: { path: LinkedTablePath }) {
+    const { control } = useFormContext<EditCdcSinkTaskFormData>();
+
+    return (
+        <div>
+            <div className="grid mb-3">
+                <FormGroup className="g-col-6" marginClass="m-0">
+                    <FormLabel>Property name</FormLabel>
+                    <FormInput type="text" control={control} name={`${path}.propertyName`} />
+                </FormGroup>
+                <FormGroup className="g-col-6" marginClass="m-0">
+                    <FormLabel>Linked collection</FormLabel>
+                    <FormInput type="text" control={control} name={`${path}.linkedCollectionName`} />
+                </FormGroup>
+                <FormGroup className="g-col-6" marginClass="m-0">
+                    <FormLabel>Source schema</FormLabel>
+                    <FormInput type="text" control={control} name={`${path}.sourceTableSchema`} />
+                </FormGroup>
+                <FormGroup className="g-col-6" marginClass="m-0">
+                    <FormLabel>Source table</FormLabel>
+                    <FormInput type="text" control={control} name={`${path}.sourceTableName`} />
+                </FormGroup>
+            </div>
+            <EditCdcSinkTaskStringValueListEditor
+                title="Join columns"
+                addButtonLabel="Add join column"
+                path={`${path}.joinColumns`}
+            />
+        </div>
+    );
+}
